+++ bank-feeds/src/codat_bankfeeds/source_accounts.py
@@ -23,30 +23,9 @@ class SourceAccounts(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
     ) -> Optional[operations.CreateSourceAccountResponseBody]:
-        r"""Create source account
+        r"""Create single source account
 
         The _Create Source Account_ endpoint allows you to create a representation of a bank account within Codat's domain. The company can then map the source account to an existing or new target account in their accounting software.
-
-        #### Account mapping variability
-
-        The method of mapping the source account to the target account varies depending on the accounting software your company uses.
-
-        #### Mapping options:
-
-        1. **API Mapping**: Integrate the mapping journey directly into your application for a seamless user experience.
-        2. **Codat UI Mapping**: If you prefer a quicker setup, you can utilize Codat's provided user interface for mapping.
-        3. **Accounting Platform Mapping**: For some accounting software, the mapping process must be conducted within the software itself.
-
-        ### Integration-specific behaviour
-
-        | Bank Feed Integration | API Mapping | Codat UI Mapping | Accounting Platform Mapping |
-        | --------------------- | ----------- | ---------------- | --------------------------- |
-        | Xero                  | ✅          | ✅               |                             |
-        | FreeAgent             | ✅          | ✅               |                             |
-        | Oracle NetSuite       | ✅          | ✅               |                             |
-        | Exact Online (NL)     | ✅          | ✅               |                             |
-        | QuickBooks Online     |             |                  | ✅                          |
-        | Sage                  |             |                  | ✅                          |
 
         > ### Versioning
         > If you are integrating the Bank Feeds API with Codat after August 1, 2024, please use the v2 version of the API, as detailed in the schema below. For integrations completed before August 1, 2024, select the v1 version from the schema dropdown below.
@@ -162,30 +141,9 @@ class SourceAccounts(BaseSDK):
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
     ) -> Optional[operations.CreateSourceAccountResponseBody]:
-        r"""Create source account
+        r"""Create single source account
 
         The _Create Source Account_ endpoint allows you to create a representation of a bank account within Codat's domain. The company can then map the source account to an existing or new target account in their accounting software.
-
-        #### Account mapping variability
-
-        The method of mapping the source account to the target account varies depending on the accounting software your company uses.
-
-        #### Mapping options:
-
-        1. **API Mapping**: Integrate the mapping journey directly into your application for a seamless user experience.
-        2. **Codat UI Mapping**: If you prefer a quicker setup, you can utilize Codat's provided user interface for mapping.
-        3. **Accounting Platform Mapping**: For some accounting software, the mapping process must be conducted within the software itself.
-
-        ### Integration-specific behaviour
-
-        | Bank Feed Integration | API Mapping | Codat UI Mapping | Accounting Platform Mapping |
-        | --------------------- | ----------- | ---------------- | --------------------------- |
-        | Xero                  | ✅          | ✅               |                             |
-        | FreeAgent             | ✅          | ✅               |                             |
-        | Oracle NetSuite       | ✅          | ✅               |                             |
-        | Exact Online (NL)     | ✅          | ✅               |                             |
-        | QuickBooks Online     |             |                  | ✅                          |
-        | Sage                  |             |                  | ✅                          |
 
         > ### Versioning
         > If you are integrating the Bank Feeds API with Codat after August 1, 2024, please use the v2 version of the API, as detailed in the schema below. For integrations completed before August 1, 2024, select the v1 version from the schema dropdown below.
@@ -271,6 +229,258 @@ class SourceAccounts(BaseSDK):
         if utils.match_response(
             http_res,
             ["400", "401", "402", "403", "404", "429", "500", "503"],
+            "application/json",
+        ):
+            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
+            raise errors.ErrorMessage(data=data)
+        if utils.match_response(http_res, ["4XX", "5XX"], "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise errors.SDKError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = await utils.stream_to_text_async(http_res)
+        raise errors.SDKError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    def create_batch(
+        self,
+        *,
+        request: Union[
+            operations.CreateBatchSourceAccountRequest,
+            operations.CreateBatchSourceAccountRequestTypedDict,
+        ],
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+    ) -> Optional[operations.CreateBatchSourceAccountResponse]:
+        r"""Create source accounts
+
+        The _Batch create source accounts_ endpoint allows you to create multiple representations of your SMB's bank accounts within Codat's domain. The company can then map the source account to an existing or new target account in their accounting software.
+
+        > ### Versioning
+        > If you are integrating the Bank Feeds API with Codat after August 1, 2024, please use the v2 version of the API, as detailed in the schema below. For integrations completed before August 1, 2024, select the v1 version from the schema dropdown below.
+
+        :param request: The request object to send.
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+
+        if not isinstance(request, BaseModel):
+            request = utils.unmarshal(
+                request, operations.CreateBatchSourceAccountRequest
+            )
+        request = cast(operations.CreateBatchSourceAccountRequest, request)
+
+        req = self.build_request(
+            method="POST",
+            path="/companies/{companyId}/connections/{connectionId}/connectionInfo/bankFeedAccounts/batch",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=False,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                True,
+                "json",
+                Optional[operations.CreateBatchSourceAccountRequestBody],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["408", "429", "5XX"])
+
+        http_res = self.do_request(
+            hook_ctx=HookContext(
+                operation_id="create-batch-source-account",
+                oauth2_scopes=[],
+                security_source=self.sdk_configuration.security,
+            ),
+            request=req,
+            error_status_codes=[
+                "400",
+                "401",
+                "402",
+                "403",
+                "404",
+                "409",
+                "429",
+                "4XX",
+                "500",
+                "503",
+                "5XX",
+            ],
+            retry_config=retry_config,
+        )
+
+        data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text, Optional[operations.CreateBatchSourceAccountResponseBody]
+            )
+        if utils.match_response(http_res, "207", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text,
+                Optional[operations.CreateBatchSourceAccountSourceAccountsResponseBody],
+            )
+        if utils.match_response(
+            http_res,
+            ["400", "401", "402", "403", "404", "409", "429", "500", "503"],
+            "application/json",
+        ):
+            data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)
+            raise errors.ErrorMessage(data=data)
+        if utils.match_response(http_res, ["4XX", "5XX"], "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise errors.SDKError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = utils.stream_to_text(http_res)
+        raise errors.SDKError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    async def create_batch_async(
+        self,
+        *,
+        request: Union[
+            operations.CreateBatchSourceAccountRequest,
+            operations.CreateBatchSourceAccountRequestTypedDict,
+        ],
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+    ) -> Optional[operations.CreateBatchSourceAccountResponse]:
+        r"""Create source accounts
+
+        The _Batch create source accounts_ endpoint allows you to create multiple representations of your SMB's bank accounts within Codat's domain. The company can then map the source account to an existing or new target account in their accounting software.
+
+        > ### Versioning
+        > If you are integrating the Bank Feeds API with Codat after August 1, 2024, please use the v2 version of the API, as detailed in the schema below. For integrations completed before August 1, 2024, select the v1 version from the schema dropdown below.
+
+        :param request: The request object to send.
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+
+        if not isinstance(request, BaseModel):
+            request = utils.unmarshal(
+                request, operations.CreateBatchSourceAccountRequest
+            )
+        request = cast(operations.CreateBatchSourceAccountRequest, request)
+
+        req = self.build_request_async(
+            method="POST",
+            path="/companies/{companyId}/connections/{connectionId}/connectionInfo/bankFeedAccounts/batch",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=False,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                True,
+                "json",
+                Optional[operations.CreateBatchSourceAccountRequestBody],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["408", "429", "5XX"])
+
+        http_res = await self.do_request_async(
+            hook_ctx=HookContext(
+                operation_id="create-batch-source-account",
+                oauth2_scopes=[],
+                security_source=self.sdk_configuration.security,
+            ),
+            request=req,
+            error_status_codes=[
+                "400",
+                "401",
+                "402",
+                "403",
+                "404",
+                "409",
+                "429",
+                "4XX",
+                "500",
+                "503",
+                "5XX",
+            ],
+            retry_config=retry_config,
+        )
+
+        data: Any = None
+        if utils.match_response(http_res, "200", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text, Optional[operations.CreateBatchSourceAccountResponseBody]
+            )
+        if utils.match_response(http_res, "207", "application/json"):
+            return utils.unmarshal_json(
+                http_res.text,
+                Optional[operations.CreateBatchSourceAccountSourceAccountsResponseBody],
+            )
+        if utils.match_response(
+            http_res,
+            ["400", "401", "402", "403", "404", "409", "429", "500", "503"],
             "application/json",
         ):
             data = utils.unmarshal_json(http_res.text, errors.ErrorMessageData)